My_form_styling
This is my FrontendMasters Initial First project Submit Where we created a form with a HTML, CSS and JavaScript also adding an error checking and validation which also include an alert message when a user places an order or leaving the input field blank. when the order are being sent it will direct be pushed to the POST method Email provided by the company.